backend/requirements.txt
fastapi==0.115.12
psycopg2-binary==2.9.11
python-dotenv==1.1.0
pytest==8.3.5
uvicorn==0.34.2
